ref(flags): Refactor LaunchDarkly integration to reusable functions
Moves core logic into re-usable functions so that we can re-use for other integrations.
